% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{appliquer_etape}
\alias{appliquer_etape}
\title{Applique une seule étape de transformation à un data.frame}
\usage{
appliquer_etape(df, etape)
}
\arguments{
\item{df}{data.frame en entrée}

\item{etape}{liste décrivant l'opération (type + paramètres)}
}
\description{
Applique une seule étape de transformation à un data.frame
}
