% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{detecter_outliers}
\alias{detecter_outliers}
\title{Détecter les outliers (IQR ou Z-score)}
\usage{
detecter_outliers(df, colonne, methode = "iqr", seuil = NULL)
}
\description{
Détecter les outliers (IQR ou Z-score)
}
