% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hygie.R
\name{hygie}
\alias{hygie}
\title{Lancer l'interface Hygie}
\usage{
hygie(port = 3838, browser = TRUE)
}
\arguments{
\item{port}{Entier. Port à utiliser (défaut : 3838)}

\item{browser}{Logical. Ouvrir dans le navigateur ? (défaut : TRUE)}
}
\value{
Lance l'application Shiny (invisiblement)
}
\description{
Lance l'application Shiny Hygie pour le traitement interactif des données.
Aucune connaissance de R n'est requise.
}
\examples{
\dontrun{
  hygie()
}
}
