% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{importer_fichier}
\alias{importer_fichier}
\title{Importer un fichier selon son extension}
\usage{
importer_fichier(chemin, options = list())
}
\description{
Importer un fichier selon son extension
}
