% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{rejouer_etapes}
\alias{rejouer_etapes}
\title{Rejoue une liste d'étapes à partir des données brutes originales}
\usage{
rejouer_etapes(donnees_originales, etapes)
}
\arguments{
\item{donnees_originales}{data.frame de départ (jamais modifié)}

\item{etapes}{liste d'étapes}
}
\value{
data.frame résultant
}
\description{
Rejoue une liste d'étapes à partir des données brutes originales
}
