% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_hygie}
\alias{run_hygie}
\title{Lancer l'application Hygie}
\usage{
run_hygie(port = 3838, launch.browser = TRUE)
}
\arguments{
\item{port}{Port sur lequel lancer l'application (par défaut 3838)}

\item{launch.browser}{Ouvrir automatiquement le navigateur (TRUE par défaut)}
}
\description{
Lancer l'application Hygie
}
